coleslaw/requirements.txt
pyyaml==6.0.1
Django==4.2.13
django-cors-headers==4.3.1
gunicorn==22.0.0
Pillow==10.3.0